refactor(marketplace): split command package and refresh maintainer docs
